requirements.txt

numpy==1.26.4
openai==1.44.0
pillow==10.4.0
pydub==0.25.1
python-dotenv==1.0.1
